LPD-26955 Move logic from StaticPagesPage.ts to reuse existent PagesAdminPage.ts
